requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gunicorn>=20.1.0
django-oauth-toolkit>=1.3.0
drf-base64